Try to handle device creation and startup

+++ Lib/Sources/UIUnitTestCLI/InstallCommand.swift
@@ -14,45 +14,54 @@ struct InstallCommand: AsyncParsableCommand {
     var osVersion: String = "16.2"
 
     mutating func run() async throws {
-        let cloneDeviceLists: String = await executeShellCommand("xcrun simctl --set testing list")
         
-        let versionsRegex = try! Regex("-- iOS \(osVersion.replacingOccurrences(of: ".", with: "\\.")) --\\n([\\n\\sA-Za-z\\d\\(\\)-]*)\\n--")
+        let timeInterval: TimeInterval = 120
+        let beginTime = Date()
         
-        guard let devices = try? versionsRegex.firstMatch(in: cloneDeviceLists)?.output[0] else {
-            print("Error regex")
-            return
+        while Date().timeIntervalSince(beginTime) < timeInterval {
+            let cloneDeviceLists: String = await executeShellCommand("xcrun simctl --set testing list")
+            
+            let versionsRegex = try! Regex("-- iOS \(osVersion.replacingOccurrences(of: ".", with: "\\.")) --\\n([\\n\\sA-Za-z\\d\\(\\)-]*)\\n--")
+            
+            guard let devices = try? versionsRegex.firstMatch(in: cloneDeviceLists)?.output[0] else {
+                print("Error regex")
+                return
+            }
+            
+            let allDevices = String(cloneDeviceLists[devices.range!])
+            
+            print(allDevices)
+            
+            let deviceRegex = try! Regex("Clone ([0-9]*) of \(deviceName) \\(([0-9A-F-]*)\\) \\(Booted\\)")
+            for match in allDevices.matches(of: deviceRegex) {
+                let deviceID = Int(String(allDevices[match.output[1].range!])) ?? 0
+                let deviceIdentifier = String(allDevices[match.output[2].range!])
+                
+                let result: String = await executeShellCommand("xcrun simctl --set testing listapps \(deviceIdentifier)")
+                
+                let appInstalled = result.contains("bruno.mazzo.ServerUITests.xctrunner")
+                //
+                if !appInstalled || forceInstall {
+                    let serverRunnerZip = Bundle.module.url(forResource: "PreBuild", withExtension: ".zip")!
+                    
+                    let tempDirectory = getTempFolder()
+                    
+                    let testRunnerZip = copyFile(file: serverRunnerZip, toFolder: tempDirectory)
+                    
+                    await executeShellCommand("unzip -o \(testRunnerZip.path) -d \(tempDirectory.relativePath)")
+                    
+                    let rootFolder = String(tempDirectory.pathComponents.joined(separator: "/").dropFirst())
+                    
+                    await executeShellCommand("xcrun simctl --set testing install \(deviceIdentifier) \(rootFolder)/ServerUITests-Runner.app")
+                }
+                
+                if await !self.isServerRunning(for: deviceID) {
+                    await executeShellCommand("xcrun simctl --set testing launch \(deviceIdentifier) bruno.mazzo.ServerUITests.xctrunner")
+                }
+            }
+            
+            try await Task.sleep(nanoseconds: 500_000_000)
         }
-        
-        let allDevices = String(cloneDeviceLists[devices.range!])
-        
-        print(allDevices)
-        
-        let deviceRegex = try! Regex("Clone [0-9]* of \(deviceName) \\(([0-9A-F-]*)\\) \\(Booted\\)")
-        for match in allDevices.matches(of: deviceRegex) {
-            let deviceIdentifier = String(allDevices[match.output[1].range!])
-            
-            let result: String = await executeShellCommand("xcrun simctl --set testing listapps \(deviceIdentifier)")
-            
-//            let appInstalled = result.contains("bruno.mazzo.ServerUITests.xctrunner")
-//
-//            if !appInstalled || forceInstall {
-                let serverRunnerZip = Bundle.module.url(forResource: "PreBuild", withExtension: ".zip")!
-                
-                let tempDirectory = getTempFolder()
-                
-                let testRunnerZip = copyFile(file: serverRunnerZip, toFolder: tempDirectory)
-                
-                await executeShellCommand("unzip -o \(testRunnerZip.path) -d \(tempDirectory.relativePath)")
-                
-                let rootFolder = String(tempDirectory.pathComponents.joined(separator: "/").dropFirst())
-                
-                await executeShellCommand("xcrun simctl --set testing install \(deviceIdentifier) \(rootFolder)/ServerUITests-Runner.app")
-//            }
-            
-            await executeShellCommand("xcrun simctl --set testing launch \(deviceIdentifier) bruno.mazzo.ServerUITests.xctrunner")
-        }
-        
-        
     }
     
     func getTempFolder() -> URL {
@@ -76,5 +85,15 @@ struct InstallCommand: AsyncParsableCommand {
         
         try! FileManager.default.copyItem(at: initialPath, to: newPath)
         return newPath
+    }
+    
+    func isServerRunning(for deviceId: Int) async -> Bool {
+        let defaultPort = 22087
+        do {
+            let _ = try await URLSession.shared.data(for: URLRequest(url: URL(string: "http://localhost:\(defaultPort + deviceId)/alive")!))
+            return true
+        } catch {
+            return false
+        }
     }
 }
